fix: - Move MainWindow to a higher directory. - Moved the processing of MainWindow to ShellPage and changed to insert ShellPage into the Content of MainWindow. - Changed namespace from block scope to file scope. - Add ViewModel.
