+++ WinuiAppDemo/Services/ClockService.cs
@@ -4,14 +4,13 @@ using NLog;
 
 using WinuiAppDemo.Services.Interfaces;
 
-namespace WinuiAppDemo.Services
-{
-    /// <inheritdoc />
-    internal class ClockService : IClockService
-    {
-        private static readonly Logger _logger = LogManager.GetCurrentClassLogger();
+namespace WinuiAppDemo.Services;
 
-        /// <inheritdoc />
-        public DateTime GetCurrentTime() => DateTime.Now;
-    }
+/// <inheritdoc />
+internal class ClockService : IClockService
+{
+    private static readonly Logger _logger = LogManager.GetCurrentClassLogger();
+
+    /// <inheritdoc />
+    public DateTime GetCurrentTime() => DateTime.Now;
 }
