+++ WinuiAppDemo/Views/ShellPage.xaml
@@ -1,20 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<Window
-    x:Class="WinuiAppDemo.Views.MainWindow"
+<Page
+    x:Class="WinuiAppDemo.Views.ShellPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:WinuiAppDemo"
+    xmlns:local="using:WinuiAppDemo.Views"
     xmlns:models="using:WinuiAppDemo.Models"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Title="WinuiAppDemo">
+    mc:Ignorable="d">
 
-    <Window.SystemBackdrop>
-        <MicaBackdrop />
-    </Window.SystemBackdrop>
-
-    <Grid x:Name="RootGrid">
+    <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="auto" />
             <RowDefinition Height="*" />
@@ -78,4 +73,4 @@
         </NavigationView>
 
     </Grid>
-</Window>
+</Page>
