+++ WinuiAppDemo/Views/UserPage.xaml.cs
@@ -2,21 +2,30 @@ using Microsoft.UI.Xaml.Controls;
 
 using NLog;
 
-namespace WinuiAppDemo.Views
-{
-    /// <summary>
-    /// Provides functionality for the settings page of the application.
-    /// </summary>
-    public sealed partial class UserPage : Page
-    {
-        private static readonly Logger _logger = LogManager.GetCurrentClassLogger();
+using WinuiAppDemo.ViewModels;
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="UserPage"/> class.
-        /// </summary>
-        public UserPage()
-        {
-            InitializeComponent();
-        }
+namespace WinuiAppDemo.Views;
+
+/// <summary>
+/// Provides functionality for the user page of the application.
+/// </summary>
+public sealed partial class UserPage : Page
+{
+    private static readonly Logger _logger = LogManager.GetCurrentClassLogger();
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="UserPage"/> class.
+    /// </summary>
+    public UserPage()
+    {
+        InitializeComponent();
+
+        ViewModel = App.GetService<UserViewModel>();
+        ViewModel.Startup();
     }
+
+    /// <summary>
+    /// Gets the view model.
+    /// </summary>
+    public UserViewModel ViewModel { get; private set; }
 }
