+++ WinuiAppDemo/appsettings.json
@@ -1,0 +1,9 @@
+{
+    "AppSettings": {
+        "LogMinLevel": "Trace",
+        "LogMaxLevel": "Fatal",
+        "LogArchiveEvery": 3,
+        "LogMaxArchiveFiles": 7,
+        "UserSettingsFileNmae": "UserSettings.json"
+    }
+}
