+++ WinuiAppDemo/userSettings.json
@@ -1,0 +1,4 @@
+﻿{
+  "Theme": 2,
+  "TimeFormat": "HH:mm:ss"
+}
